refactor trace
