+++ lib/components/index.ts
@@ -25,7 +25,7 @@ export { SilkscreenPath } from "./primitive-components/SilkscreenPath"
 export { SilkscreenRect } from "./primitive-components/SilkscreenRect"
 export { SilkscreenText } from "./primitive-components/SilkscreenText"
 export { SmtPad } from "./primitive-components/SmtPad"
-export { Trace } from "./primitive-components/Trace"
+export { Trace } from "./primitive-components/Trace/Trace"
 export { TraceHint } from "./primitive-components/TraceHint"
 export { Via } from "./primitive-components/Via"
 export { Battery } from "./normal-components/Battery"
