+++ lib/components/normal-components/Capacitor.ts
@@ -7,7 +7,7 @@ import {
 } from "lib/utils/constants"
 import { NormalComponent } from "../base-components/NormalComponent"
 import { Port } from "../primitive-components/Port"
-import { Trace } from "../primitive-components/Trace"
+import { Trace } from "../primitive-components/Trace/Trace"
 import { formatSiUnit } from "format-si-unit"
 
 export class Capacitor extends NormalComponent<
